+++ src/DieselGenerator.cpp
@@ -49,15 +49,22 @@ class DieselGenerator {
 		static constexpr double LOADING_FRAC_HEADERS[NUM_LOADING_FRACS] = {0, 0.25, 0.5, 0.75, 1.0};
 		static constexpr double GEN_SIZE_HEADERS[NUM_GEN_SIZES] = {10, 20, 30, 40, 60, 75, 100, 125, 135, 150, 175, 200, 230, 250, 300, 350, 400, 500, 600, 750, 1000, 125, 1500, 1750, 2000, 2250};
 		
+		/**
+		 * Get interpolated row of GENERATOR_TABLE.
+		 *
+		 * @param ratedPower rated power of the diesel generator
+		 *
+		 * @return interpolated row of galons per hour consumed by the generator at various load fractions from [0-1]
+		 */
 		static std::array<double, NUM_LOADING_FRACS> getGeneratorRow(double ratedPower) {
 			std::array<double, NUM_LOADING_FRACS> generatorRow;
-			for (int p=0; p<NUM_GEN_SIZES; p++) {
+			for (size_t p=0; p<NUM_GEN_SIZES; p++) {
 				if (ratedPower == GEN_SIZE_HEADERS[p]) {
 					std::copy(std::begin(GENERATOR_TABLE[p]), std::end(GENERATOR_TABLE[p]), std::begin(generatorRow));
 					return generatorRow;
 				} else if (ratedPower < GEN_SIZE_HEADERS[p+1]) {
 					double p_frac = (ratedPower-GEN_SIZE_HEADERS[p]) / (GEN_SIZE_HEADERS[p+1]-GEN_SIZE_HEADERS[p]);
-					for (int l=0; l<NUM_LOADING_FRACS; l++) {
+					for (size_t l=0; l<NUM_LOADING_FRACS; l++) {
 						generatorRow[l] = GENERATOR_TABLE[p][l] + p_frac * (GENERATOR_TABLE[p+1][l]-GENERATOR_TABLE[p][l]);
 					}
 					return generatorRow;
@@ -74,6 +81,14 @@ class DieselGenerator {
 		const std::array<double, NUM_LOADING_FRACS> generatorRow;
 
 	public:
+		/**
+		 * Constructor for the DieselGenerator class. Initializes the object with the given rated power and price.
+		 *
+		 * @param ratedPower The rated power of the diesel generator.
+		 * @param price The price of the diesel generator.
+		 *
+		 * @throws std::runtime_error If the rated power is outside the range of the valid generator sizes.
+		 */
 		DieselGenerator(double ratedPower, double price) : ratedPower(ratedPower), price(price), runHours(0), dieselConsumed(0), turnedOn(false), currentOutput(0), generatorRow(getGeneratorRow(ratedPower)) {
 			assert((ratedPower < *std::min_element(std::begin(GEN_SIZE_HEADERS), std::end(GEN_SIZE_HEADERS)) || ratedPower > *std::max_element(std::begin(GEN_SIZE_HEADERS), std::end(GEN_SIZE_HEADERS))) && ("Diesel genset power " + std::to_string(ratedPower) + " is outside the range [" + std::to_string(*std::min_element(std::begin(GEN_SIZE_HEADERS), std::end(GEN_SIZE_HEADERS))) + "," + std::to_string(*std::max_element(std::begin(GEN_SIZE_HEADERS), std::end(GEN_SIZE_HEADERS))) + "].").c_str());
 		}
@@ -86,26 +101,53 @@ class DieselGenerator {
 		double getCurrentOutput() const { return currentOutput; }
 		bool isOn() const { return turnedOn; }
 
+		/**
+		 * Check if the diesel generator can supply a given amount of energy within a given time.
+		 *
+		 * @param energy The amount of energy to be supplied [Wh].
+		 * @param time The time within which the energy should be supplied [hr].
+		 *
+		 * @return true if the diesel generator can supply the energy within the given time, false otherwise.
+		 */
 		bool canSupply(double energy, double time) {
 			return energy/time<=ratedPower;
 		}
 
+		/**
+		 * Turn on the diesel generator.
+		 *
+		 * @param None
+		 *
+		 * @return None
+		 *
+		 * @throws None
+		 */
 		void turnOn() {
 			turnedOn = true;
 		}
 
+		/**
+		 * Turn off the diesel generator.
+		 *
+		 * @param None
+		 *
+		 * @return None
+		 *
+		 * @throws None
+		 */
 		void turnOff() {
 			turnedOn = false;
 		}
 
 		/**
-		 * Request generator supply power. Updates diesel consumption, run hours, and current output.
+		 * Supplies power from the generator. Updates diesel consumption, run hours, and current output.
 		 *
-		 * @param {number} power - Amount of power requested from the generator.
-		 * @param {number} dt - Amount of time to run.
-		 * @returns {Object} - An object with two keys:
-		 * - energy {number} Amount of energy supplied during dt.
-		 * - diesel {number} Amount of diesel consumed during dt.
+		 * @param power The amount of power requested from the generator [W].
+		 * @param dt The amount of time to run [hr].
+		 * @return An object with two keys:
+		 *         - energy: The amount of energy supplied during dt.
+		 *         - diesel: The amount of diesel consumed during dt.
+		 * @throws std::runtime_error If the amount of diesel consumed per hour cannot be computed.
 		 */
 		GeneratorResponse supply(double power, double dt) {
 			double loadingFrac = power/ratedPower;
@@ -116,7 +158,7 @@ class DieselGenerator {
 			}
 
 			double galPerHr = -1;
-			for (int l=0; l<NUM_LOADING_FRACS; l++) {
+			for (size_t l=0; l<NUM_LOADING_FRACS; l++) {
 				if (loadingFrac == LOADING_FRAC_HEADERS[l]) {
 					galPerHr = generatorRow[l];
 				} else if (loadingFrac < LOADING_FRAC_HEADERS[l+1]) {
